+++ src/pages/EventDetail.tsx
@@ -383,7 +383,7 @@ export default function EventDetail() {
   ];
 
   return (
-    <div className="flex flex-col gap-4 pb-8">
+    <div className="flex-1 min-h-0 overflow-y-auto overscroll-contain flex flex-col gap-4 pb-8">
       {/* Back button */}
       <button
         onClick={() => navigate("/events")}